feat(feed): POST /drop — the phone upload route
phone-screenshot-drop U6. The image is the raw request body and the
caption rides the query: multipart would mean a boundary-scanning state
machine (or a new dependency) inside fly's most security-sensitive
listener to carry one file whose metadata fits in a query string, and a
header can't hold a newline or an emoji without encoding anyway.

Every refusal routes through one drain_and_refuse helper, and this is
load-bearing rather than tidy. Verified in tiny_http-0.12.0's
equal_reader.rs: an unread body is drained at drop time via
`vec![0; remaining_to_read]`, sized from the client-declared
Content-Length the crate parses with no cap. Responding early doesn't
skip the upload, it defers it and sizes a buffer from a number the
client chose. Because EqualReader::read decrements that counter,
draining first leaves the drop path nothing to allocate. The residual
is documented rather than papered over: draining trades an allocation
for a socket read, and a declared length past DRAIN_LIMIT still leaves
a remainder.

Two departures from the plan, both from checking rather than assuming:

- The format sniff runs before the mid-stream cap check, not after. It
  needs only 16 bytes, so going first means a non-image never creates a
  file — the plan's own "nothing written before a decision that doesn't
  need the bytes". Consequence: an oversize non-image reports badFormat.
- The planned test "a body exceeding the cap despite a small declared
  Content-Length" is unsatisfiable. Under Content-Length framing
  tiny_http bounds the reader at the declared length, so the route can
  never see more than was declared and the early check is total. The
  mid-stream take(cap + 1) bound exists for CHUNKED bodies, which
  declare no length at all — the test now exercises that, which is the
  case that actually reaches it.

Delivery goes through one seam rather than exposed steps because the
ordering is load-bearing: the image must be published after every
refusal check and before any text reaches the pane, or either a
refusal leaves residue or the agent is told to read a path that hasn't
been renamed into place. DropConfig groups the four new parameters so
start() doesn't grow to twelve arguments.

+++ src-tauri/src/feed/wire.rs
@@ -74,7 +74,7 @@ pub struct AgentEntry {
     /// Additive and nullable like `lastReplyAt`/`questionPendingAt` before it:
     /// `default` keeps both an older stored payload and an older consumer valid.
     #[serde(default)]
-    pub pane_id: Option<u32>,
+    pub pane_id: Option<u64>,
 }
 
 /// One selectable option of a pending question, as the wire carries it
